requirements.txt
numpy>=1.25.0
torch>=2.0.0
torchvision>=0.13.1
Pillow>=9.0.1
python-telegram-bot>=20.3
